requirements.txt
flask
flask-login
requests
google-cloud-storage
google-cloud-firestore
